[Fix] 새로운 HomeCoordinator 프로토콜과 구현체 생성
