+++ MateRunner/MateRunner/Presentation/HomeScene/Coordinator/Protocol/HomeCoordinator.swift
@@ -2,15 +2,12 @@
 //  HomeCoordinator.swift
 //  MateRunner
 //
-//  Created by 전여훈 on 2021/11/09.
+//  Created by 전여훈 on 2021/11/10.
 //
 
-import UIKit
+import Foundation
 
-protocol SettingCoordinator: Coordinator {
-    func createHomeViewController() -> UIViewController
-    func pushRunningModeSettingViewController()
-    func pushMateRunningModeSettingViewController(with settingData: RunningSetting?)
-    func pushDistanceSettingViewController(with settingData: RunningSetting?)
-    func pushRunningPreparationViewController(with settingData: RunningSetting?)
+protocol HomeCoordinator: Coordinator {
+    func showSettingFlow()
+    func showRunningFlow()
 }
